+++ src/main/java/com/osudpotro/posmaster/purchase/purchasecart/PurchaseCartItemAddRequest.java
@@ -1,11 +1,13 @@
 package com.osudpotro.posmaster.purchase.purchasecart;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PurchaseCartItemAddRequest {
     private Long productId;
     private Long productDetailId;
-    private Double purchasePrice;
-    private Long purchaseQty;
-    private Double purchaseLinePrice;
+    private BigDecimal purchasePrice;
+    private Integer purchaseQty;
+    private BigDecimal purchaseLinePrice;
 }
